requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
sentence-transformers
faiss-cpu
openai
python-multipart